[fea] redesign - Navigation component (partly completed) (#1509)
This PR add new navigation component - atm. it is not included in dashboard / header, therefor it is not impact dashboard in any way. It is partly completed, however it is the state that we can start preparing design. Following steps are: place component into new header, design it (Primoz) and refactor used services (e.g. zemNavigationService).

New library: angular-vs-repeat - quite ok VS for navigation component needs.