[fea] redesign header breadchrumb + new navigation service (#1528)
This PR adds redesigned header breadcrumb along with new navigation service zemNavigationNewService, to be used by new components (navigation, breadcrumb component). The service itself depends on the older one zemNavigationService, but provides clean access for new components. It is first iteration of this service and will be enhanced in the future with idea to completely separate it from older one, with its own endpoint service and cache.