add sentry through index.html (#2075)
This PR adds Sentry/Raven configuration to frontend/client.
Integrated through index.html.
This was the simplest way for integration, and probably could be done cleaner, but I couldn't find simple solution to do it. There is option to do it with bower, but I had problems with minimization. If there is a need we can switch later.